+++ arch/arm/armv7m/include/kickos/arch/context.h
@@ -57,7 +57,7 @@ struct arch_context
     // well as one inside the user stack, because a thread preempted mid-dispatch is running
     // there. Read at F_CTX_KERNEL_SP in switch.S. Zero for a TCB outside the pool, which
     // reaches neither site.
-    uint32_t kernel_sp;
+    uintptr_t kernel_sp;
 
 #if defined(KICKOS_TELEMETRY) && KICKOS_TELEMETRY
     // Owning thread's trace id (stamped once in thread_create). switch.S reads it
